fix(db,api): resolve 4 pre-existing test failures across repos and migrations
- Cast SUM(file_size_bytes)::BIGINT in trash_repo purge_preview to fix
  sqlx decode error (PostgreSQL SUM(bigint) returns numeric)
- Switch trash_api filtered-by-type test from ID-based negative check to
  count assertion (BIGSERIAL sequences across tables can collide)
- Exempt append-only tables (events, notifications) from updated_at
  requirement in schema_conventions timestamp test
- Add missing FK index on notification_preferences(event_type_id) to
  satisfy test_all_fks_have_indexes convention check

All 176 tests now pass.

+++ apps/backend/crates/db/src/repositories/trash_repo.rs
@@ -131,8 +131,9 @@ impl TrashRepo {
         }
 
         // Estimate bytes from scene_video_versions (has file_size_bytes column).
+        // Note: SUM(bigint) returns numeric in PostgreSQL, so cast back to bigint.
         let bytes: (i64,) = sqlx::query_as(
-            "SELECT COALESCE(SUM(file_size_bytes), 0) \
+            "SELECT COALESCE(SUM(file_size_bytes)::BIGINT, 0) \
              FROM scene_video_versions WHERE deleted_at IS NOT NULL",
         )
         .fetch_one(pool)
